refactor(tanstack-start): expose withPayload as the sole root export
Matches `@payloadcms/next`, where the package root exports only `withPayload`.

- Rename `src/vite` to `src/withPayload`.
- Root barrel exports only `withPayload` (drop the option helpers and types,
  which had no external consumers).
- Remove the `/vite` subpath export; import `withPayload` from the root instead.
