--- packages/tanstack-start/src/exports/vite.ts
@@ -1,13 +1,0 @@
-export {
-  payloadReactOptions,
-  payloadRscOptions,
-  payloadTanstackStartOptions,
-  withPayload,
-} from '../vite/index.js'
-export type {
-  PayloadPluginOptions,
-  PayloadTanstackStartOptionsArgs,
-  WithPayloadBuilder,
-  WithPayloadBuilderContext,
-  WithPayloadOptions,
-} from '../vite/index.js'
