+++ packages/tanstack-start/src/withPayload/config/external.ts
