+++ packages/tanstack-start/src/withPayload/workarounds/stripDistStyleImports.ts
