+++ test/vite.tanstack.config.ts
@@ -1,4 +1,4 @@
-import { withPayload } from '@payloadcms/tanstack-start/vite'
+import { withPayload } from '@payloadcms/tanstack-start'
 import { tanstackStart } from '@tanstack/react-start/plugin/vite'
 import viteReact from '@vitejs/plugin-react'
 import rsc from '@vitejs/plugin-rsc'
